+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Usage:
 * Fix #336: Fix pull from insecure registries and pull registry authentication for JIB
 * Fix #332: OpenShift build resources are deleted (as long as build config manifest is available)
 * Fix #350: Prevents default docker configuration overwriting XML assembly configuration
+* Fix #340: Exclude the main artifact from Docker build when Fat Jar is detected (JavaExecGenerator)
 
 ### 1.0.0-rc-1 (2020-07-23)
 * Fix #252: Replace Quarkus Native Base image with ubi-minimal (same as in `Dockerfile.native`)